Assistant checkpoint: Fix uint256 to uint8 type conversion error
Assistant generated file changes:
- contracts/NFTVaultV3.sol: Fix type mismatch in getMaxLTV function

---

User prompt:

~/workspace$ npx hardhat test
TypeError: Type uint256 is not implicitly convertible to expected type uint8.
   --> contracts/NFTVaultV3.sol:291:46:
    |
291 |         RiskModel storage model = riskModels[riskTier];
    |                                              ^^^^^^^^

Error HH600: Compilation failed

For more info go to https://hardhat.org/HH600 or run Hardhat with --show-stack-traces

+++ contracts/NFTVaultV3.sol
@@ -287,7 +287,7 @@ contract NFTVaultV3 is Ownable, ReentrancyGuard {
     }
 
     function getMaxLTV(address collection, uint256 tokenId) public view returns (uint256) {
-        uint256 riskTier = collectionRiskTiers[collection];
+        uint8 riskTier = collectionRiskTiers[collection];
         RiskModel storage model = riskModels[riskTier];
         uint256 baseLTV = model.baseLTV;
